Update actions.js
Proposed a change to line 541. Currently, the aws-lex-web-ui default error message is 'Sorry, I was unable to process your message. Try again later.' which does not provide any hint of what the exact issue is. I have found this error message pop up either when someone has been inactive for quite some time (timed out) or someone has not logged in. I propose to change the wording to provide a more intuitive explanation of why user is receiving the error message and provide some concrete steps to be able to successfully use the chatbot.